loading model from file:
Base class:
because parse_value is a virtual function and is needed when loading a file, file cant be loaded in base class constructor, as such, that code was offloaded to a function.

model factory:
new function to load model from file (uses base class's one).
new constructor to build  a model factory from a file.